refactor: Extract enum to separate file
