+++ index-handlers/src/main/java/no/sikt/nva/nvi/index/query/CandidateQuery.java
@@ -2,7 +2,6 @@ package no.sikt.nva.nvi.index.query;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.emptySet;
-import static java.util.Objects.isNull;
 import static java.util.Objects.nonNull;
 import static no.sikt.nva.nvi.common.utils.JsonUtils.jsonPathOf;
 import static no.sikt.nva.nvi.index.model.document.ApprovalStatus.APPROVED;
@@ -38,9 +37,7 @@ import static no.sikt.nva.nvi.index.utils.SearchConstants.TITLE;
 import static no.sikt.nva.nvi.index.utils.SearchConstants.TYPE;
 import static no.sikt.nva.nvi.index.utils.SearchConstants.YEAR;
 
-import com.fasterxml.jackson.annotation.JsonValue;
 import java.time.ZonedDateTime;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
@@ -245,41 +242,6 @@ public record CandidateQuery(
     return globalStatuses.stream()
         .map(QueryFunctions::globalStatusQuery)
         .reduce(QueryFunctions::matchAtLeastOne);
-  }
-
-  public enum QueryFilterType {
-    COLLABORATION("collaboration"),
-    OTHERS_APPROVE("approvedByOthers"),
-    OTHERS_REJECT("rejectedByOthers"),
-    NEW_AGG("pending"),
-    NEW_COLLABORATION_AGG("pendingCollaboration"),
-    PENDING_AGG("assigned"),
-    PENDING_COLLABORATION_AGG("assignedCollaboration"),
-    APPROVED_AGG("approved"),
-    APPROVED_COLLABORATION_AGG("approvedCollaboration"),
-    REJECTED_AGG("rejected"),
-    REJECTED_COLLABORATION_AGG("rejectedCollaboration"),
-    DISPUTED_AGG("dispute"),
-    ASSIGNMENTS_AGG("assignments"),
-    EMPTY_FILTER("");
-
-    private final String filter;
-
-    QueryFilterType(String filter) {
-      this.filter = filter;
-    }
-
-    public static Optional<QueryFilterType> parse(String candidate) {
-      var testValue = isNull(candidate) ? "" : candidate;
-      return Arrays.stream(values())
-          .filter(item -> item.getFilter().equalsIgnoreCase(testValue))
-          .findAny();
-    }
-
-    @JsonValue
-    public String getFilter() {
-      return filter;
-    }
   }
 
   public static class Builder {
