+++ index-handlers/src/main/java/no/sikt/nva/nvi/index/utils/SearchConstants.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import java.util.Optional;
 import no.sikt.nva.nvi.index.model.search.CandidateSearchParameters;
 import no.sikt.nva.nvi.index.query.CandidateQuery;
-import no.sikt.nva.nvi.index.query.CandidateQuery.QueryFilterType;
+import no.sikt.nva.nvi.index.query.QueryFilterType;
 import nva.commons.core.Environment;
 import org.opensearch.client.opensearch._types.mapping.DateProperty;
 import org.opensearch.client.opensearch._types.mapping.DoubleNumberProperty;
